requirements.txt
pytesseract
Pillow
pyperclip
opencv-python
numpy
